тесты с mocha для es2015 заработали
