--- jest.config.js
@@ -1,4 +1,0 @@
-export default {
-  testMatch: ['**/__tests__/functions/*.test.js'],
-  transform: {},
-};
