+++ package.json
@@ -17,13 +17,19 @@
     "eslint": "8.42.0",
     "eslint-config-htmlacademy": "9.1.1",
     "esm": "^3.2.25",
-    "jest": "^29.5.0"
+    "jasmine": "^5.0.2",
+    "jest": "^29.5.0",
+    "karma": "^6.4.2",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-jasmine": "^5.1.0",
+    "mocha": "^10.2.0",
+    "mocha-esm": "^1.1.1"
   },
   "license": "MIT",
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/",
-    "test": "jest --transformIgnorePatterns '/node_modules/(?!@babel/runtime)'"
+    "test": "mocha --require mocha-esm test.mjs"
   },
   "engines": {
     "node": "18",
